fantasy-draft/pyproject.toml
[project]
name = "football-project"
version = "0.1.0"
requires-python = ">= 3.13.3"
dependencies = [
    "aesara>=2.9.3",
    "arviz>=0.22.0",
    "bambi>=0.15.0",
    "datetime>=5.5",
    "great-tables>=0.18.0",
    "ipykernel>=6.30.1",
    "jax>=0.7.1",
    "marimo>=0.14.17",
    "matplotlib>=3.10.5",
    "nfl-data-py>=0.3.2",
    "numpy==2.2",
    "nutpie>=0.15.2",
    "polars>=1.32.2",
    "preliz>=0.20.0",
    "pyarrow>=21.0.0",
    "pyjanitor>=0.31.0",
    "pymc>=5.25.1",
    "requests>=2.32.4",
    "ruff>=0.12.7",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "selenium>=4.35.0",
    "stats>=0.1.2a0",
    "xarray>=2025.8.0",
]
